Create p-value_Rsquare_LogisticRegression.py
It is always good to know the p-values and Rsqaure value of the machine learning model. 
Predictors having p-values greater than 0.05 (95% confidence interval) can be dropped from the model as they are statistically insignificant. And with the value of Rsquare, we can determine that how much variation in y variable can be explained by the x variable (predictors) in the model.